task: Move functions from arweave.ts to other files
