+++ src/gql.ts
@@ -2,7 +2,7 @@ import * as common from './common';
 import * as types from './types/base_Types';
 import * as arfsTypes from './types/arfs_Types';
 import * as gqlTypes from './types/gql_Types';
-import { getTransactionData } from './public/arweave';
+import { getTransactionData } from './gateway';
 import { deriveDriveKey, driveDecrypt } from './crypto';
 import Arweave from 'arweave';
